pyproject.toml
[tool.poetry]
name = "protocol-benchmark"
version = "0.1.0"
description = "Protocol benchmark"
authors = ["Andrzej Bogdanowicz <bahdanovich@gmail.com>"]

packages = [
    { include="data_gen", from="." },
    { include="grpc_betterproto", from="." },
    { include="graphql_strawberry", from="." },
]

[tool.poetry.dependencies]
python = ">=3.8,<4"
betterproto = {extras = ["compiler"], version = "^1.2.5"}
Faker = "^5.0.1"
orjson = "^3.4.3"
tqdm = "^4.51.0"
strawberry-graphql = "^0.44.2"
uvicorn = "^0.13.1"
requests = "^2.25.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
grpc-stubs = "^1.24.2"
grpcio = "^1.33.2"
grpcio-tools = "^1.33.2"
pylint = "^2.6.0"
mypy = "^0.790"
pytest = "^6.1.2"

[tool.poetry.scripts]
grpc-betterproto-server = 'grpc_betterproto.grpc_server:run'
grpc-betterproto-client = 'grpc_betterproto.grpc_client:run'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
